整理项目内容
